Generalized CPM-SC-FDMA Transmission Scheme Suitable for Future Aeronautical Telemetry using MATLAB
n this project, we propose CPM-SC-FDMA, a novel transmission scheme
which combines many of the key characteristics of continuous phase
modulation (CPM) with Single Carrier Frequency Division Multiple Access
(SC-FDMA) in order to produce a power efficient, robust modulation
scheme which is suitable for high-rate multiple-access aeronautical
telemetry applications. The basis of our approach is found in the
observation that the discrete-time samples from a CPM waveform
constitute a constant envelope time domain sequence which can be
DFT-pre-coded and subsequently mapped to a set of orthogonal subcarriers
for an FDMA-style transmission. By a judicious selection of the
subcarrier mapping scheme, we show that this new hybrid waveform retains
the high power efficiency of CPM. Because it is based on CPM, it is
compatible with existing telemetry signaling formats. In addition, it is
characterized by a high spectral efficiency due to the concept of
overlapping but mutually orthogonal narrowband subcarriers, thus
allowing non-interfering multiple-access.